* Fix MSVC2015 build warnings

+++ src/Core/Utility/Helper.h
@@ -100,7 +100,7 @@ template <typename T>
 struct hash : std::unary_function<T, size_t> {
 	std::size_t operator()(T const& matrix) const {
 		size_t seed = 0;
-		for (size_t i = 0; i < matrix.size(); i++) {
+		for (int i = 0; i < (int)matrix.size(); i++) {
 			auto elem = *(matrix.data() + i);
 			seed ^= std::hash<typename T::Scalar>()(elem) + 0x9e3779b9 +
 					(seed << 6) + (seed >> 2);
