+++ src/IO/FileFormat/FileLOG.cpp
@@ -101,7 +101,7 @@ bool WritePinholeCameraTrajectoryToLOG(const std::string &filename,
 	}
 	for (size_t i = 0; i < trajectory.extrinsic_.size(); i++ ) {
 		const auto &trans = trajectory.extrinsic_[i];
-		fprintf(f, "%lu %lu %lu\n", i, i, i + 1);
+		fprintf(f, "%d %d %d\n", (int)i, (int)i, (int)i + 1);
 		fprintf(f, "%.8f %.8f %.8f %.8f\n", trans(0,0), trans(0,1), trans(0,2),
 				trans(0,3) );
 		fprintf(f, "%.8f %.8f %.8f %.8f\n", trans(1,0), trans(1,1), trans(1,2),
