+++ src/Test/TestCameraPoseTrajectory.cpp
@@ -55,7 +55,7 @@ int main(int argc, char *argv[])
 	std::vector<std::shared_ptr<const Geometry>> pcds;
 	for (size_t i = 0; i < trajectory.extrinsic_.size(); i++) {
 		char buff[DEFAULT_IO_BUFFER_SIZE];
-		sprintf(buff, "%scloud_bin_%lu.pcd", argv[2], i);
+		sprintf(buff, "%scloud_bin_%d.pcd", argv[2], (int)i);
 		if (filesystem::FileExists(buff)) {
 			auto pcd = CreatePointCloudFromFile(buff);
 			pcd->Transform(trajectory.extrinsic_[i]);
